flow-ingest/pyproject.toml
[project]
name = "ztna-flow-ingest"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "redis==5.1.0",
  "loguru==0.7.2",
  "pyyaml==6.0.2",
  "uvloop==0.20.0",
]

[project.optional-dependencies]
test = [
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "fakeredis==2.25.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
